requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
requests
openpyxl
sqlalchemy
chromadb
PyMuPDF
python-multipart